Compose based Entity and Components now have message handlers for setting properties.
Entities with components section and initial values for properties
propertly send set_property messages to the components after they are
created and added to the Entity.

Entities initialize their data in constructor.

Components cannot do this since they don't have block storage yet during
the constructor.

To fix this, I've added a #init_data message which gets sent to a
component immediately after it is added to an entity, and initialization
for Component data (properties and fields) happens in this message handler.