Added gaen initialization code.
Turns out we appear to need to do drawing on main thread
in osx.  At least we don't have the flexibility we do on
windows in terms of where to do the drawing. I need to go
back to the windows side and see if I can re-design how
rendering is initiated each frame and get it to happen
on the main thread, then osx and I think ios porting will
become much simpler.