Fixed osx startup issue
Renderer was not declared as a pointer property in our view.

Don't know why I didn't see this before, but what can I say, it stumped
me. On windows it can be a local since we control the message pump, on
OSX, we need a pointer to maintain the renderer around outside the view
initialization handler.  Super big Doh!

New issue is shaders don't compile since version 330 is unsupported.
Joy.